refactor: move gpu data type related extension methods and tests to their own files
