+++ Velaptor/OpenGL/Buffers/LineGPUBuffer.cs
@@ -10,6 +10,7 @@ using System.Linq;
 using Batching;
 using Carbonate.UniDirectional;
 using Exceptions;
+using ExtensionMethods;
 using Factories;
 using GPUData;
 using NativeInterop.OpenGL;
